Add files via upload
Credit, Grocery, and Joint csvs are provided to show how the program will work.  A categories CSV will be created in the same directory as the convert.py file, this keeps track of the store names the user provides and what I call the vendor codes that represent those store names.  After the program has executed, a Purchases csv will be formed that has all of the purchases consolidated and formatted, and then an excel of that Purchases csv will open.  This can then be implemented into an excel budget of your choosing.